20 html/js finish

+++ Listas/Lista-01/04-Operadores_Matematicos_e_Logicos/19-operadores_matematicos_e_logicos.js
@@ -7,7 +7,7 @@ const resultDiv = document.getElementById("resultDiv");
 // Ouvinte evento button
 
 procButton.addEventListener("click", () => {
-  const numQ = parseFloat(numQualquer.value);
+  const numQ = parseInt(numQualquer.value);
 
   // verificar validade
   if (isNaN(numQ)) {
